Fixed timer for when the volume is too loud
If the volume is above -5 then the volume cannot get much louder
so switch to a timer as long as the volume stays that low

It was taking one timed picture than taking a bunch of pictures
based on the volume and fixed that. Had to invalidate the timers
differently.